# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TÖPP API",
    "description": "REST API интернет-магазина ковров; цены в целых рублях. /docs для интерактивного OpenAPI.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/carts": {
      "post": {
        "summary": "Create Cart",
        "operationId": "create_cart_api_v1_carts_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/checkout/options": {
      "get": {
        "summary": "Options",
        "operationId": "options_api_v1_checkout_options_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/checkout/quote": {
      "post": {
        "summary": "Quote",
        "operationId": "quote_api_v1_checkout_quote_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteIn"
              },
              "example": {
                "address": {
                  "city": "Москва",
                  "street": "Лесная",
                  "house": "1"
                },
                "deliveryMethod": "manual_delivery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cartBearer": []
          }
        ]
      }
    },
    "/api/v1/orders": {
      "post": {
        "summary": "Create Order",
        "operationId": "create_order_api_v1_orders_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderIn"
              },
              "example": {
                "customerName": "Иван",
                "phone": "+79991234567",
                "address": {
                  "city": "Москва",
                  "street": "Лесная",
                  "house": "1"
                },
                "deliveryMethod": "manual_delivery",
                "paymentMethod": "on_delivery",
                "personalDataConsent": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "example": {
                  "orderId": "00000000-0000-4000-8000-000000000000",
                  "orderNumber": "f9126af6-bced-439a-a050-a77d9c1c570b",
                  "status": "new",
                  "currency": "RUB",
                  "subtotal": 12990,
                  "delivery": 0,
                  "total": 12990,
                  "deliveryMethod": "manual_delivery",
                  "paymentMethod": "on_delivery",
                  "items": [
                    {
                      "variantId": "00000000-0000-4000-8000-000000000000",
                      "name": "Лунна",
                      "sku": "SKU-001",
                      "sizeLabel": "160 × 230 см",
                      "unitPrice": 12990,
                      "quantity": 1,
                      "lineTotal": 12990
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт цен, остатков или условий",
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CART_CHANGED",
                    "message": "Цены или остатки изменились",
                    "details": {
                      "items": [
                        {
                          "variantId": "00000000-0000-4000-8000-000000000000",
                          "available": 0,
                          "expectedPrice": 12990,
                          "currentPrice": 13990
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cartBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "properties": {
          "city": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "City"
          },
          "street": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Street"
          },
          "house": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "House"
          },
          "apartment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Apartment"
          },
          "postalCode": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postalcode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "city",
          "street",
          "house"
        ],
        "title": "Address"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderIn": {
        "properties": {
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "deliveryMethod": {
            "type": "string",
            "const": "manual_delivery",
            "title": "Deliverymethod"
          },
          "customerName": {
            "type": "string",
            "maxLength": 150,
            "minLength": 2,
            "title": "Customername"
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+?[0-9 ()-]{10,25}$",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "on_delivery",
              "manual_confirmation"
            ],
            "title": "Paymentmethod"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "personalDataConsent": {
            "type": "boolean",
            "const": true,
            "title": "Personaldataconsent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "address",
          "deliveryMethod",
          "customerName",
          "phone",
          "paymentMethod",
          "personalDataConsent"
        ],
        "title": "OrderIn"
      },
...
      "QuoteIn": {
        "properties": {
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "deliveryMethod": {
            "type": "string",
            "const": "manual_delivery",
            "title": "Deliverymethod"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "address",
          "deliveryMethod"
        ],
        "title": "QuoteIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "cartBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Токен из POST /api/v1/carts"
      },
      "adminBasic": {
        "type": "http",
        "scheme": "basic",
        "description": "ADMIN_USERNAME и пароль администратора"
      }
    }
  }
}